remove v8 update and refactored code
