+++ includes/classes/class-installation.php
@@ -45,9 +45,6 @@ class ATBDP_Installation {
 			'directorist_7123_remove_upload_files_cap',
 			'directorist_7123_update_db_version',
 		],
-		'8.0.0' => [
-			'directorist_800_update_db_version',
-		]
 	);
 
 	/**
